# perf(sim): hard-wire tuned smoothing constants (`MAX_SAMPLES=8`, `αG=0.392`, `αA=0.127`)