This patch permanently **replaces the old heuristic defaults** with the values produced by the ε-greedy tuner, baking them directly into the simulation module:

| Constant | Old value(s) | **New (tuned) value** | Role |
|----------|--------------|-----------------------|------|
| `MAX_SAMPLES`    | 5 (implicit)            | **8**     | Length of position history used to compute angular velocity |
| `SMOOTH_ALPHA_G` | ~0.30 – 0.40 runtime-derived | **0.392** | Exponential-smoothing α when **on ground** |
| `SMOOTH_ALPHA_A` | 0.05 – 0.20 runtime-derived  | **0.127** | Exponential-smoothing α when **in air** |

#### What changed
* **No runtime knobs** – the constants are now literal numbers at the top of `simulation/player.lua`  
  ```lua
  local MAX_SAMPLES      = 8       -- tuned window size
  local SMOOTH_ALPHA_G   = 0.392   -- tuned ground α
  local SMOOTH_ALPHA_A   = 0.127   -- tuned air α